backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
joblib
scikit-learn
xgboost
python-multipart
python-jose
python-dotenv
shap
geopandas